move logic from services to sagas
