+++ App/Sagas/likesPost/index.spec.js
@@ -5,9 +5,9 @@ import { call, select, put, take, fork } from 'redux-saga/effects';
 
 import {channelStoreMock, channelsStoreWithKeyMock, likeWithKeyMock} from '../../../Tests/mock/';
 import { likedChannelsActions, channelsActions, likedChannelsTypes } from '../../Redux/';
-import { likesPostToFirebase } from '../../Services';
 import {uidSelector, likedChannelsSelector} from '../selector';
 import { likesPostIncrease, mergeLikedChannelToLocal, likeOnServer } from '../likesPost';
+import {likesPostToFirebase} from './firebase';
 
 test.serial.group('Normal: increase', () => {
   const generator = likesPostIncrease({channel: channelStoreMock()});
